pyproject.toml
[project]
name = "flowergarden"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26",
    "matplotlib>=3.8",
    "streamlit>=1.32",
    "supabase>=2.0",
    "plotly>=5.0",
    "streamlit-image-coordinates>=0.1.0",
    "Pillow>=10.0",
]
